Update lc0.Dockerfile to enable Leela's preload option.
### Extended Description:
This change addresses an issue where the engine fails to start thinking immediately when playing as the second player, particularly after the opening book phase. The observed behaviour in the GUI is as follows:

1. The backend waits for the first engine to play its move after the opening book is completed.
2. When the first engine responds with `readyok`, the timer for the first player starts, and the first player makes its move.
3. At this point, the second player (Leela in this case) is expected to start thinking. However, since Leela has not already preloaded its state, it first needs to initialize itself before it can begin thinking, leading in time loss.

To address this, the `preload` option is enabled, this ensures that Leela loads its state and prepares itself before being asked to think.